Try harder to avoid a change feed rewind after a shard move
In the previous attempt [1] we improved the logic by spawning workers on the
matching target shards only. However, that wasn't enough as workers might still
reject the passed in sequence from the old node when it asserts ownership
locally on each shard.

Re-use the already existing replacement clause, where after uuid is matched, we
try harder to find the highest viable sequence. To use the unit test setup as
an example, if the shard moved from node1 to node2, and recorded epoch `{node2,
10}` on the new node, then a sequence generated on node1 before the move, for
example 12, would rewind down to 10 only when calculated on new location on
node2, instead of being rewound all the way down to 0.

[1] https://github.com/apache/couchdb/commit/e83935c7f8c3e47b47f07f22ece327f6529d4da0