Minimize rewinds when a node is down (#3792)
Our existing logic for handling rewinds in the changes feed addresses
the following cases:

- A node that contributed to a sequence is in maintenance mode
- A shard that contributed to a sequence has been split

This patch adds support for cases where the node that contributed to a
client-supplied sequence is down at the beginning of the request
handling. It reuses the same logic as the maintenance mode case as these
two situations really ought to be handled the same way.

A future improvement would be to unify the "node down" and "shard split"
logic so that we could handle the compound case, e.g. replacing a shard
from a down node with a pair of shards from nodes that cover the same
range.

Fixes #3788

Co-authored-by: Nick Vatamaniuc <vatamane@gmail.com>