Consolidate ExUnit integration test configuration
Two main pieces of work here:

1) Consoidate elixir / elixir-only / elixir-suite into two targets:

- elixir-integration, which only runs ExUnit integration tests
- elixir-suite, which depends on elixir-integration and also calls
  format + credo

This preserves the flexibility of the previous targets while reducing
code duplication and removing some obsolete configurations

2) Limit the number of places where we define test configurations

Previously we were relying on dev/run to inject config entries in
addition to specifiying a .ini file on the command-line. We sometimes
also write config settings in the test definitions themselves. This is
at least an initial attempt to just define the config settings in one
place.