Handle `not_found` lookups removing ddoc cache key
For non-existent ddocs, a new ddoc_cache_entry is spawned for each
call to ddoc_cache:open. Multiple calls to open the same non-existent
ddoc will create multiple cache entries with the same Key but
different PIDs. This can result in khash lookups returning `not_found`
as in the error below:

[error] 2021-12-10T02:25:21.622743Z node1@127.0.0.1 <0.18923.9> -------- gen_server ddoc_cache_lru terminated with reason: no match of right hand value not_found at ddoc_cache_lru:remove_key/2(line:308) <= ddoc_cache_lru:handle_info/2(line:219) <= gen_server:try_dispatch/4(line:637) <= gen_server:handle_msg/6(line:711) <= proc_lib:init_p_do_apply/3(line:249)

This checks the return values of `khash:lookup` and only proceeds to
delete keys if the results are other than `not_found`.