Index on a "virtual" field computed from normal fields
This puts in just enough machinery to support indexing on a function of
a document's content inside Mango, without going through the JS engine.
Say we have a document like:

    {
      "bar": "a b c"
    }

If we want to index on the individual "words" of `bar`, i.e. find this
doc via the keys "a", "b" or "c", we'd normally need a JS map function:

    function (doc) {
      for (let word of doc.bar.split(" ")) {
        emit(word)
      }
    }

This patch lets us do this inside Mango by defining an index on a
"virtual" field whose value is a function of the doc's other fields. We
put a view containing this in a design doc:

    "map": {
      "fields": {
        "bar_words": {
          "$explode": { "$field": "bar", "$separator": " " }
        }
      }
    }

And this lets us perform `_find` queries for e.g. { "bar_words": "b" }
to get our original document.

As this is a proof of concept designed to get the index machinery
working, `$explode` is the only function defined.