NIF bindings for jq
Here we add a module named `couch_jq` which exports an interface to the
jq [1] library, letting us run jq programs against CouchDB docs held in
memory as Erlang terms. The exported functions are:

- `couch_jq:compile/1`: takes a binary containing a jq expression, and
  compiles it. Returns a resource object holding the resulting struct
  that stored the parsed expression.

- `couch_jq:eval/2`: takes a compiled jq program and a JSON document and
  returns the results of evaluating the program against the doc, as a
  list.

Both functions return either `{ok, Result}` or `{error, Message}`.

This assumes that jq programs will in general return multiple results.
Because the Erlang list is built from the tail, the results are returned
in the reverse order that they're emitted by jq. We're going to put
these values into an index where they'll get re-sorted so the order is
not important.

Most of the native code here is concerned with translating between
Erlang and jq representations of JSON values. Erlang terms are checked
to make sure they conform to the Couch document representation, and
anything else is rejected.

[1]: https://stedolan.github.io/jq/