Allow multiple jq expressions in index definitions
If have a document like:

    {
      "foo": "a b",
      "bar": "x y"
    }

And an index definition like:

    [
      { "foo_words": { "$jq": '.foo | split(" ") | .[]' } },
      { "bar_words": { "$jq": '.bar | split(" ") | .[]' } }
    ]

Then this should produce four index keys for the document:

    ("a", "x")
    ("a", "y")
    ("b", "x")
    ("b", "y")

This lets us query on multiple virtual fields in a single query. The
implementation here allows jq expressions (that return multiple values)
to be mixed with normal field access that returns a single value;
`flatten_keys/1` returns the product of any multi-valued index fields.
For example, above `foo_words` produces values `["a", "b"]` and
`bar_words` produces `["x", "y"]`, and we multiply this out giving the
four keys above.