Compile jq expressions once per index, not once per document
The first implementation of building Mango indexes from jq expressions
re-compiled the jq program for every document it processed. This is
wasteful and we can instead to this once, when the index definitions are
updated.

When the index definition is loaded in, field descriptions of the form
`{[{<<"$jq">>, Expr}]}` are converted to `{jq, CompiledExpr}`. This
non-JSON term is used because the compiled jq program can be held in
memory but not serialised back out into the design doc.