use undefined for timed out responses
Under heavy load, fabric_update_doc workers will return timeout via
rexi. Therefore no reponses will be populate the response dictionary.
This leads to badargs when we do dict:fetch for keys that do not exist.
This checks for the existence of the key and then returns undefined
if it does not exist, which aligns with the same code path of when we
the number of docs < 100.