Improve index building during shard splitting
Previously we didn't check responses from get_state/2 or await/2 functions when
building indices. If an index updater crashed, and the index never finished
building, the get_state/2 call would simply return an error and the process
would exit normally. Then, the shard splitting job would count that as a
success and continue to make progress.

To fix that, make sure to check the response to all the supported indexing
types and wait until they return an `ok` result.

Additionally, increase the index building resilience to allow for more retries
on failure, and for configurable retries for individual index builders.