demonstrate native encryption
Key management is _not_ done and so this scheme is not secure.

The AES_MASTER_KEY value should be retrieved interactively and need
not be the same for different databases.

Overwriting the first 40 bytes of the file with any other value
renders the file unreadable.

We use AES in Counter Mode, which ensures we can encrypt and decrypt
any section of the file without padding or alignment. The ciphertext
is the same length as the plaintext. This mode provides
confidentiality but not authentication.