Integrate raft algorithm (WIP)
couch_raft.erl is a complete implementation of the raft algorithm but
currently only manages an in-memory state machine and log.

Preliminary work is also here to add a new btree inside the `.couch`
files, which will be the real raft log. The intent is that log entries
can be removed from this log and applied to by_id and by_seq trees
atomically.

raft log is not preserved over compaction yet because reading the
compactor code hurts my eyes.

Anyway, it's progress and hopefully we're going somewhere cool.