Optimize couch_util:reorder_results/2,3
This function is used in the hot path of _revs_diff and _bulk_docs API calls.
Those could always use a bit more optimization:

  * In `_revs_diff` it's used when fetching all the FDIs to see which docs are
    missing in `couch_btree:lookup/2`.

  * In `_bulk_docs` it's used in the `fabric_doc_update` when finalizing the
    response.

Using erlperf in #4051 noticed an at most 5x speedup from using a map instead
of a dict. Since a map already falls back to a proplist for small sizes, skip
the length guard.

Some erlperf examples from #4051:

500 Keys
```
> f(Keys), f(Res), {Keys, Res} = Gen(500), ok.

> erlperf:run(#{runner => {couch_util, reorder_results2, [Keys, Res, 100, dict]}}).
2407
> erlperf:run(#{runner => {couch_util, reorder_results2, [Keys, Res, 100, map]}}).
11639
```

Using a map without the guard, which is the change in this this PR:

```
> f(Keys), f(Res), {Keys, Res} = Gen(500), ok.
ok

> erlperf:run(#{runner => {couch_util, reorder_results, [Keys, Res]}}).
12395
> erlperf:run(#{runner => {couch_util, reorder_results, [Keys, Res]}}).
12508
```

As a bonus this also cleans up the code a bit, too.