Add two larger couch_key_tree:find_missing/2 tests
These were used to generate the random rev trees used for bechmarking. But they
also assert useful properties for the function so let's add them to test_util.

Since we added a shuffle/1 function to test_util, DRY some extra shuffle/1
functions as we seem to have a bunch of them sprinkled here and there.

Also, since the rev tree generation code was changed, need to adjust the
`excessive_memory_when_stemming` test as well. To do it, reverted the commit
indicated in the test and then adjusted parameters until the test failed. Then
reset the revert and it failed.