Fix flaky "validate doc update" elixir test
We saw it fail quite a few times already so adding a retry around it.

Unfortunately it could now fail with a timeout instead of an assert but since
we're updating a document there is no idempotent way to testing it (retry
until, then do the actual assert again below).