separate follower and candidate timeouts
From;

ARC: Analysis of Raft Consensus - 4.2

"As the authors use the same timer range for candidates and followers,
in Figure 4.1 we are waiting a minimum of 150ms (and up to twice that)
before restarting an election, despite the fact that, on average, a
node receives all of its responses within 15ms"

We separate the timeouts and set the candidate timeout smaller than
the follower timeout. In a contested election (where multiple
candidates each gain a minority of votes) we should elect a leader
faster than otherwise.