Update couch_replicator_use_checkpoints_tests
Use the clustered version of source and target endpoints.

Also, use common setup and teardown functions. The test was previously quite
clever in trying to save an extra few lines by parameterizing the "use
checkpoing" vs "don't use checkpoints" scenarios with a foreachx and a custom
notifier function. Instead, opt for more clarity and use the usual TDEF_FE
macro and just setup two separate test one which uses checkpoints and one which
doesn't.

Another major win is is using the utility db comparison function instead of a
duplicate copy of it.