Statistically skip _revs_diff in the replicator
When the revisions are consistently missing from the target, for
example when replicating to a new target database, it's wasteful to
keep calling _revs_diff.  Reuse the same algorithm used when skipping
calls to _bulk_get if there are too many attachments to statistically
start skipping calls to _revs_diff.

Also use the same record for keeping track of stats so rename
`bulk_get_stats` to `fetch_stats`.

The update and attempt functions are re-used as well, there are just
separate wrappers which pass in different values for the threshold and
other parameters.