Create 3.2.2-docs branch after migrating docs to main repo
This branch contains the 3.2.2 version of the docs but migrated to the
main CouchDB repo. After switching the URL repo in RDT we want to make
sure we still display the stable (aliased to 3.2.2) docs in RDT not
just the latest ones.

Main integration PR https://github.com/apache/couchdb/pull/4164