Upgrade to latest Sphinx 5.1.1 and fix the top Edit on Github link
"Edit on Github" link wasn't working and pointing to a non-existent "index"
file. This was most likely because we override the index.html template, so the
source renderer doesn't know what to link it to. Instead opt to make the top
level "Edit on Github" point the docs readme page. Individual doc pages will
point to their respective source pages.

In addition, we were quite a bit behind on on the sphinx version 1.5 vs 5.1.1
this has caused quite a bit of headaches over the years as we had to pin
jinja2, docutils and other libraries versions to avoid breaking things.

The main issue after the upgrade was that the http domain became a bit more
strict. After vendoring it in, we had to make a few multipart http examples use
plaintext. On the positive side, it found broken references in the admin guide,
so those were fixed. Since we're using the latest 5.1.1 opt to just use a
python3 venv with a short setup.sh script. That should make it easier for
contributors to build docs locally.