Fix eventsource change feed
At some point when the `stop` callback arg got a 3rd `Pending`
element, we forgot to update the eventsource callback. As a result, we
ended up in a catch-all clause and we appended an unmatched ']' and a
json `"last_seq"` element.

While at it, clean up of the similar upgrade clauses in other callback
instances.