Remove the long deprecated bigcouch 0.4 change sequence support
We always ran a regex match for it, then failed, and tried the regular pattern.
Regexes are pretty fast (15-30 microseconds on my laptop), but it's still a
waste of resources. For comparison a `config:get/3` is about microsecond, so
it's like needlessly performing 10 extra config lookup per `_changes` feed
request.

Improve unit tests by having more cases and one test per case.

Remove unreachable `"now"` clause in `validate_start_seq/2`. By that point
`get_start_seq/2` should have transformed `"now"` into a proper sequence so
it's confusing validating that case as passing. At that point it should be
considered a failure, which is what one of the new unit tests checks for.