Explicitly maintain a fully connected cluster
Previously, it was possible for the nodes to disconnect, and for that state to
persist util the nodes restarted. Some fabric requests could reconnect the
nodes as a side-effect of sending remote messages, but most of the fabric
requests currently start a rexi monitor, which immediately delivers a
`rexi_DOWN` message to the coordinator for worker nodes not in the `[node() |
nodes()]` list. That happens before `erlang:send/2,3` gets called, so there is
nothing there to eventually reconnect the nodes.

To avoid relying on the random request reconnecting the cluster, use an
explicit monitor process. It does the initial connections, as well as
periodically maintains them.

[1] https://github.com/apache/couchdb/blob/main/src/rexi/src/rexi_monitor.erl#L45