Expose the batched open_doc_revs function variant.
This is the batched version of the existing `open_doc_revs/4` function. It
turns out the internal implementation of `open_doc_revs/4` already operated on
batches, just that the batch size had always been 1, so here we just expose the
batched variant.

One subtle difference from the single variant, is there are options which apply
to the whole request, and some apply to individual documents only. In the
previous `open_doc_revs/4` both were intermingled in the single Options
argument. In the batched version, the per-document options are passed in each
`{{Id, Rev}, DocOpts}` tuple. Here DocOpts will usually have `{atts_since,
...}` option, but it could have others, like `ejson_body`, for example.

The API is also exposed in `fabric_rpc. This function will be the used to
optimize `_bulk_get` as described in issue 4183. To allow incremental node
updates, let's have the backend implementation in a separate commit so it could
be included in separate rolling updates releases.

Add some tests both for the original `open_doc_revs/4` function and batched
variant. Since the internal implementation is exactly the same, the batched
variant mainly checks how multiple results are returned and that per document
options are applied correctly.

Issue: https://github.com/apache/couchdb/issues/4183