Avoid refresh messages piling up in prometheus server
Previously, the cast `refresh` message didn't explicitly cancel the previous
timer. If a `gen_server:call(Server, refresh)` was made and a `refresh` message
was already in the message queue, it was possible to end up with two parallel
refresh periods (cycles): one with the regular refresh period, the other
generated by the regular refresh gen_server call.

In addition, drain all the previous refresh messages before scheduling a new
timer. This should further help lower the chance of refresh cycles piling up on
top of each other.

The fix is mostly theoretical as technically I don't think anything triggers a
`refresh` gen_server:call/2 calls in practice. It would have to be a manual or
an externally scripted call.