Implement global password hasher process (#4240)
Implement a global password hasher process. The new behavior reduces the 
hashing calls from 2 * N (N equals the number of `couch_server` processes)
down to 2 calls. The first call is triggered by the change in the config file and
the second call through `config:set` to write the hashed result back into the
config file. If we want to reduce this to one call only, we need to implement
some more intelligence into the config part, to prevent the triggers for such calls. 

The password hasher is implemented as a `gen_server` and started with the
`couch_primary_services` supervisior.

Fixes #4236.