Ensure prevent_overlapping_partitions stays false in Erlang 25+
It's already false in 23 and 24 but will start to be enabled in 25+. We don't
rely on global for process registration any more, and have our own
auto-connection module. So we don't want to be caught by surprise in Erlang 25+
since there is some additional coordination and resource usage needed when this
option is true. See https://github.com/erlang/otp/issues/6470 for an example.