revert jiffy to 1.0.9
jiffy > 1.0.9 breaks the windows builds. We tracked this down to jiffy
getting rebar3 support in version 1.1.0 and fixes to that support in
1.1.1. We shipped our last release with jiffy 1.0.9 which directly
preceded 1.1.0 (tho our tag is CouchDB-1.0.9). As far as I can tell
there are no substantial fixes* in 1.1.0+ so we might as well unblock
our release by reverting to jiffy 1.0.9.

* https://github.com/davisp/jiffy/compare/1.0.9...1.1.1